pyproject.toml
[project]
name = "shipyard"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk==0.1.59",
    "click==8.2.1",
]

[project.scripts]
shipyard = "shipyard.cli:main"

[project.optional-dependencies]
dev = [
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "pyyaml==6.0.2",
    "ruff==0.13.0",
    "pre-commit==4.5.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["shipyard*"]

[tool.setuptools.package-data]
shipyard = ["templates/*.yml", "prompts/*.md", "py.typed"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=shipyard --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
branch = true
source = ["shipyard"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
